add backend topup

+++ AdopPix.Models/Auction.cs
@@ -16,8 +16,8 @@ namespace AdopPix.Models
         public int HourId { get; set; }
         public DateTime StartTime { get; set; }
         public DateTime StopTime { get; set; }
-        public double OpeningPrice { get; set; }
-        public double HotClose { get; set; }
+        public decimal OpeningPrice { get; set; }
+        public decimal HotClose { get; set; }
         public string Description { get; set; }
         public DateTime Created { get; set; }
