+++ AdopPix.Models/UserProfile.cs
@@ -16,10 +16,10 @@ namespace AdopPix.Models
         public string Fname { get; set; }
         public string Lname { get; set; }
         public string Description { get; set; }
-        public double Money { get; set; }
+        public decimal Money { get; set; }
         public DateTime BirthDate { get; set; }
-        public int Point { get; set; }
-        public int Rank { get; set; }
+        public decimal Point { get; set; }
+        public decimal Rank { get; set; }
         public DateTime Created { get; set; }
 
         public User User { get; set; }
